docs: rewrite examples and README to the gi: ESM import style
Convert all examples to ES modules using the new direct-import syntax
('import Gtk from "gi:Gtk-4.0"', node-gtk's own API from 'node-gtk'), renaming
them .js -> .mjs and adapting to ESM main-loop semantics (run() returns
immediately; no top-level 'return'). Each carries a 'node --import
node-gtk/register' run line. The ts-demo now uses gi: imports too (typed via the
generated shim). README leads with the ESM style as the default and moves the
CommonJS form to its own section at the end.
