+++ examples/cairo-font.mjs
@@ -1,12 +1,16 @@
 /*
- * cairo-font.js
+ * cairo-font.mjs
+ *
+ * Run with:  node --import node-gtk/register examples/cairo-font.mjs
  */
 
-const fs = require('fs')
-const path = require('path')
-const ft = require('freetype2')
-const gi = require('../')
-const Cairo = gi.require('cairo')
+import Cairo from 'gi:cairo'
+import fs from 'node:fs'
+import path from 'node:path'
+import { fileURLToPath } from 'node:url'
+import ft from 'freetype2'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 fs.readFile(path.join(__dirname, './Hack-Regular.ttf'), function(err, buffer) {
   if (err)
@@ -29,5 +33,5 @@ function loadFont(buffer) {
   font.setSynthesize(Cairo.FtSynthesize.BOLD)
   const synth = font.getSynthesize()
   console.log({ font, synth, bold: 1 })
-  console.assert(synth === Cairo.FtSynthesize.BOLD) 
+  console.assert(synth === Cairo.FtSynthesize.BOLD)
 }
