+++ examples/hello-world.mjs
@@ -1,11 +1,13 @@
 /*
  * A minimal "Hello World" using libadwaita (GTK4-based).
+ *
+ * Run with:  node --import node-gtk/register examples/hello-world.mjs
  */
 
-const gi = require('../lib/');
-const GLib = gi.require('GLib', '2.0');
-const Gtk = gi.require('Gtk', '4.0');
-const Adw = gi.require('Adw', '1');
+import { startLoop } from 'node-gtk';
+import GLib from 'gi:GLib-2.0';
+import Gtk from 'gi:Gtk-4.0';
+import Adw from 'gi:Adw-1';
 
 const loop = GLib.MainLoop.new(null, false);
 const app = new Adw.Application('com.github.romgrk.node-gtk.hello', 0);
@@ -22,8 +24,10 @@ app.on('activate', () => {
   window.on('close-request', () => (loop.quit(), app.quit(), false));
   window.present();
 
-  gi.startLoop();
+  startLoop();
   loop.run();
 });
 
-process.exit(app.run([]));
+// Under ESM the run call returns immediately; the app keeps running and the
+// process exits when the window is closed (loop.quit()/app.quit() above).
+app.run([]);
