+++ examples/test-application.mjs
@@ -1,7 +1,7 @@
-const gi = require('.');
-const Gtk = gi.require('Gtk', '4.0');
-const Gio = gi.require('Gio', '2.0');
-const GLib = gi.require('GLib', '2.0');
+import { startLoop } from 'node-gtk';
+import Gtk from 'gi:Gtk-4.0';
+import Gio from 'gi:Gio-2.0';
+import GLib from 'gi:GLib-2.0';
 
 class TestApplication extends Gtk.Application {
   constructor() {
@@ -19,7 +19,7 @@ class TestApplication extends Gtk.Application {
       process.exit(0);
     });
 
-    gi.startLoop();
+    startLoop();
     this.loop = GLib.MainLoop.new(null, false);
     this.loop.run();
   }
